refactor: rebase 후 이미지 업로드 API 파일 위치 이동 - #247
