+++ product/src/main/java/com/back/product/adapter/in/web/api/ImageApiController.java
@@ -1,4 +1,4 @@
-package com.back.product.adapter.in;
+package com.back.product.adapter.in.web.api;
 
 import com.back.common.response.CommonResponse;
 import com.back.image.dto.enums.ImageCategory;
