+++ product/src/main/java/com/back/product/adapter/in/web/controller/ApiV1ImageController.java
@@ -1,10 +1,11 @@
-package com.back.product.adapter.in;
+package com.back.product.adapter.in.web.controller;
 
 import com.back.image.app.ImageFacade;
 import com.back.image.dto.enums.ImageCategory;
 import com.back.common.code.SuccessCode;
 import com.back.common.response.CommonResponse;
 import com.back.image.dto.response.ImageUploadResponseDto;
+import com.back.product.adapter.in.web.api.ImageApiController;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
